% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PollyRFunctions.R, R/RcppExports.R
\name{MicroGenotyper}
\alias{MicroGenotyper}
\title{Does something}
\usage{
MicroGenotyper(bam_vec, lookup_file, scaff_vec, output_names)

MicroGenotyper(bam_vec, lookup_file, scaff_vec, output_names)
}
\arguments{
\item{bam_vec}{something}

\item{scaff_vec}{something}

\item{output_names}{something}

\item{lookup_table}{something}
}
\description{
Does something

Does something
}
