% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PollyRFunctions.R, R/RcppExports.R
\name{Polly}
\alias{Polly}
\title{Does something}
\usage{
Polly(
  region_size,
  scaff_vector,
  SIthreshold,
  freq_name,
  bam_vec,
  lookup_file,
  MicroThreshold,
  SIOutputName,
  MicroOutputNames,
  PollyMicroOutputName,
  FinalOutputName
)

Polly(
  region_size,
  scaff_vector,
  SIthreshold,
  freq_name,
  bam_vec,
  lookup_file,
  MicroThreshold,
  SIOutputName,
  MicroOutputNames,
  PollyMicroOutputName,
  FinalOutputName
)
}
\arguments{
\item{region_size}{something}

\item{scaff_vector}{something}

\item{SIthreshold}{something}

\item{freq_name}{something}

\item{bam_vec}{something}

\item{lookup_file}{something}

\item{MicroThreshold}{something}

\item{SIOutputName}{something}

\item{MicroOutputNames}{something}

\item{PollyMicroOutputName}{something}

\item{FinalOutputName}{something}
}
\description{
Does something

Does something
}
