% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PollyRFunctions.R, R/RcppExports.R
\name{PollyMicros}
\alias{PollyMicros}
\title{Does something}
\usage{
PollyMicros(CSV_names, desired_scaffolds, threshold, output_name)

PollyMicros(CSV_names, desired_scaffolds, threshold, output_name)
}
\arguments{
\item{CSV_names}{something}

\item{desired_scaffolds}{something}

\item{threshold}{something}

\item{output_name}{something}
}
\description{
Does something

Does something
}
