% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PollyRFunctions.R, R/RcppExports.R
\name{PollySI}
\alias{PollySI}
\title{Does something}
\usage{
PollySI(freq_name, scaff_vector, threshold)

PollySI(freq_name, scaff_vector, threshold)
}
\arguments{
\item{freq_name}{The a string indicating the VCFtools .frq file. Be sure to include the ".frq" suffix. See Supplementary Materials for instructions on creation.}

\item{scaff_vector}{something}

\item{threshold}{something}
}
\description{
Does something

PollySI() outputs a matrix of SNP and indel loci at which both the major and minor allele frequency are greater than or equal to an adjustable threshold.
}
